+++ src/app/shared/PersonList.svelte
@@ -3,6 +3,8 @@
   import {uniq, pluck} from "ramda"
   import {batch} from "hurdak"
   import {Tags} from "src/util/nostr"
+  import {createScroller} from "src/util/misc"
+  import {getModal} from "src/partials/state"
   import Content from "src/partials/Content.svelte"
   import Spinner from "src/partials/Spinner.svelte"
   import PersonSummary from "src/app/shared/PersonSummary.svelte"
@@ -12,9 +14,15 @@
   export let type
   export let pubkey
 
+  let limit = 20
   let pubkeys = []
 
+  const loadMore = async () => {
+    limit += 20
+  }
+
   onMount(() => {
+    const scroller = createScroller(loadMore, {element: getModal()})
     const sub =
       type === "follows"
         ? subscribe({
@@ -38,12 +46,15 @@
             }),
           })
 
-    return () => sub.close()
+    return () => {
+      sub.close()
+      scroller.stop()
+    }
   })
 </script>
 
 <Content gap="gap-2">
-  {#each pubkeys as pubkey (pubkey)}
+  {#each pubkeys.slice(0, limit) as pubkey (pubkey)}
     <PersonSummary {pubkey} />
   {:else}
     <Spinner />